+++ mongodb/src/main/scala/quasar/physical/mongodb/WorkflowExecutor.scala
@@ -20,7 +20,7 @@ import quasar.Predef._
 import quasar.{NameGenerator => QNameGenerator}
 import quasar.connector.{EnvironmentError, EnvErrT}
 import quasar.contrib.scalaz.eitherT._
-import quasar.fp.kleisli._
+import quasar.contrib.scalaz.kleisli._
 import quasar.fs._
 import quasar.javascript._
 import quasar.physical.mongodb.execution._
